Remove id from infodict, wip
