+++ satpy/readers/__init__.py
@@ -24,23 +24,18 @@
 
 """
 
-import glob
 import logging
 import numbers
 import os
 from abc import ABCMeta, abstractmethod, abstractproperty
 from collections import namedtuple
-from datetime import datetime, timedelta
-from fnmatch import fnmatch
 
 import numpy as np
 import six
 import yaml
 
 from satpy.config import config_search_paths, glob_config, runtime_import
-from satpy.plugin_base import Plugin
-from satpy.projectable import Dataset
-from trollsift.parser import Parser, globify
+from satpy.projectable import DatasetID, DATASET_KEYS
 
 try:
     import configparser
@@ -49,119 +44,14 @@ except ImportError:
 
 LOG = logging.getLogger(__name__)
 
-DATASET_KEYS = ("name", "wavelength", "resolution", "polarization",
-                "calibration", "modifiers")
-DatasetID = namedtuple("DatasetID", " ".join(DATASET_KEYS))
-DatasetID.__new__.__defaults__ = (None, None, None, None, None, None)
-
-
-class DatasetID(DatasetID):
-    """Identifier for all `Dataset` objects.
-
-    DatasetID is a namedtuple that holds identifying and classifying
-    information about a Dataset. There are two identifying elements,
-    ``name`` and ``wavelength``. These can be used to generically refer to a
-    Dataset. The other elements of a DatasetID are meant to further
-    distinguish a Dataset from the possible variations it may have. For
-    example multiple Datasets may be called by one ``name`` but may exist
-    in multiple resolutions or with different calibrations such as "radiance"
-    and "reflectance". If an element is `None` then it is considered not
-    applicable.
-
-    A DatasetID can also be used in SatPy to query for a Dataset. This way
-    a fully qualified DatasetID can be found even if some of the DatasetID
-    elements are unknown. In this case a `None` signifies something that is
-    unknown or not applicable to the requested Dataset.
-
-    Args:
-        name (str): String identifier for the Dataset
-        wavelength (float, tuple): Single float wavelength when querying for
-                                   a Dataset. Otherwise 3-element tuple of
-                                   floats specifying the minimum, nominal,
-                                   and maximum wavelength for a Dataset.
-                                   `None` if not applicable.
-        resolution (int, float): Per data pixel/area resolution. If resolution
-                                 varies across the Dataset then nadir view
-                                 resolution is preferred. Usually this is in
-                                 meters, but for lon/lat gridded data angle
-                                 degrees may be used.
-        polarization (str): 'V' or 'H' polarizations of a microwave channel.
-                            `None` if not applicable.
-        calibration (str): String identifying the calibration level of the
-                           Dataset (ex. 'radiance', 'reflectance', etc).
-                           `None` if not applicable.
-        modifiers (tuple): Tuple of strings identifying what corrections or
-                           other modifications have been performed on this
-                           Dataset (ex. 'sunz_corrected', 'rayleigh_corrected',
-                           etc). `None` or empty tuple if not applicable.
-    """
-    @staticmethod
-    def name_match(a, b):
-        """Return if two string names are equal
-
-        Args:
-            a (str): DatasetID.name or other string
-            b (str): DatasetID.name or other string
-        """
-        return a == b
-
-    @staticmethod
-    def wavelength_match(a, b):
-        """Return if two wavelengths are equal
-
-        Args:
-            a (tuple or scalar): (min wl, nominal wl, max wl) or scalar wl
-            b (tuple or scalar): (min wl, nominal wl, max wl) or scalar wl
-        """
-        if type(a) == type(b):
-            return a == b
-        elif a is None or b is None:
-            return False
-        elif isinstance(a, (list, tuple)) and len(a) == 3:
-            return a[0] <= b <= a[2]
-        elif isinstance(b, (list, tuple)) and len(b) == 3:
-            return b[0] <= a <= b[2]
-        else:
-            raise ValueError("Can only compare wavelengths of length 1 or 3")
-
-    def __eq__(self, other):
-        if isinstance(other, str):
-            return self.name_match(self.name, other)
-        elif isinstance(other, numbers.Number) or \
-                        isinstance(other, (tuple, list)) and len(other) == 3:
-            return self.wavelength_match(self.wavelength, other)
-        else:
-            return super(DatasetID, self).__eq__(other)
-
-    def __hash__(self):
-        return tuple.__hash__(self)
-
-    @classmethod
-    def from_dict(cls, d, **kwargs):
-        args = []
-        for k in DATASET_KEYS:
-            val = kwargs.get(k, d.get(k))
-            # force modifiers to tuple
-            if k == 'modifiers' and val is not None:
-                val = tuple(val)
-            args.append(val)
-
-        return cls(*args)
-
-    def to_dict(self):
-        return dict(zip(DATASET_KEYS, self))
-
-AREA_KEYS = ("name", "resolution", "terrain_correction")
-AreaID = namedtuple("AreaID", " ".join(AREA_KEYS))
-AreaID.__new__.__defaults__ = (None, None, None, None, None)
-
 
 class MalformedConfigError(Exception):
     pass
 
 
 class DatasetDict(dict):
-    """Special dictionary object that can handle dict operations based on dataset name, wavelength, or DatasetID
+    """Special dictionary object that can handle dict operations based on
+    dataset name, wavelength, or DatasetID.
 
     Note: Internal dictionary keys are `DatasetID` objects.
     """
@@ -325,8 +215,8 @@ class DatasetDict(dict):
             d["calibration"] = key.calibration
             d["polarization"] = key.polarization
             d["modifiers"] = key.modifiers
-            d['id'] = key
-            # you can't change the wavelength of a dataset, that doesn't make sense
+            # you can't change the wavelength of a dataset, that doesn't make
+            # sense
             if "wavelength" in d and d["wavelength"] != key.wavelength:
                 raise TypeError("Can't change the wavelength of a dataset")
