+++ satpy/readers/viirs_sdr.py
@@ -295,14 +295,13 @@ class VIIRSSDRFileHandler(HDF5FileHandler):
             self.scale_swath_data(out.data, out.mask, factors)
 
         ds_info.update({
-            "name": dataset_id.name,
-            "id": dataset_id,
             "units": ds_info.get("units", file_units),
             "platform_name": self.platform_name,
             "sensor": self.sensor_name,
             "start_orbit": self.start_orbit_number,
             "end_orbit": self.end_orbit_number,
         })
+        ds_info.update(dataset_id.to_trimmed_dict())
         cls = ds_info.pop("container", Projectable)
         return cls(out, **ds_info)
