+++ satpy/tests/utils.py
@@ -49,13 +49,12 @@ def _create_fake_compositor(ds_id, prereqs, opt_prereqs):
     from satpy import Projectable
     c = mock.MagicMock()
     c.info = {
-        'id': ds_id,
         'prerequisites': tuple(prereqs),
         'optional_prerequisites': tuple(opt_prereqs),
     }
+    c.update(ds_id.to_trimmed_dict())
     c.return_value = Projectable(data=np.arange(5),
-                                 id=ds_id,
-                                 **ds_id.to_dict())
+                                 **ds_id.to_trimmed_dict())
     # c.prerequisites = tuple(prereqs)
     # c.optional_prerequisites = tuple(opt_prereqs)
     return c
@@ -68,14 +67,14 @@ def _create_fake_modifiers(name, prereqs, opt_prereqs):
 
     def _mod_loader(*args, **kwargs):
         class FakeMod(CompositeBase):
+
             def __init__(self, *args, **kwargs):
                 self.info = {}
 
             def __call__(self, datasets, optional_datasets, **info):
-                if name == 'res_change' and datasets[0].info['id'].resolution is not None:
+                if name == 'res_change' and datasets[0].id.resolution is not None:
                     i = datasets[0].info.copy()
                     i['resolution'] = i['resolution'] * 5
-                    i['id'] = DatasetID.from_dict(i)
                 else:
                     i = datasets[0].info
                 info = datasets[0].info.copy()
@@ -156,8 +155,7 @@ def _reader_load(dataset_keys):
         for ds in dataset_ids:
             if ds == k:
                 loaded_datasets[ds] = Projectable(data=np.arange(5),
-                                                  id=ds,
-                                                  **ds.to_dict())
+                                                  **ds.to_trimmed_dict())
     return loaded_datasets
